refactoring for release
