+++ app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 26
     defaultConfig {
-        applicationId "com.example.jc.eatery_android"
+        applicationId "com.cornellappdev.android.eatery"
         minSdkVersion 21
         targetSdkVersion 26
         versionCode 1
