+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.jc.eatery_android">
+    package="com.cornellappdev.android.eatery">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
@@ -14,7 +14,7 @@
         android:label="Eatery"
         android:supportsRtl="true">
         <activity
-            android:name=".MainActivity"
+            android:name="com.cornellappdev.android.eatery.MainActivity"
             android:theme="@style/AppTheme"
             android:screenOrientation="portrait"
             >
@@ -24,12 +24,12 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MenuActivity"
+            android:name="com.cornellappdev.android.eatery.MenuActivity"
             android:theme="@style/NoActionBar"
             android:screenOrientation="portrait"
             />
         <activity
-            android:name=".SplashActivity"
+            android:name="com.cornellappdev.android.eatery.SplashActivity"
             android:theme="@style/NoActionBar"
             android:screenOrientation="portrait"
             >
@@ -40,7 +40,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".WeeklyMenuActivity"
+            android:name="com.cornellappdev.android.eatery.WeeklyMenuActivity"
             android:theme="@style/NoActionBar"
             android:screenOrientation="portrait"
             />
@@ -50,7 +50,7 @@
             android:value="@string/google_maps_key" />
 
         <activity
-            android:name=".MapsActivity"
+            android:name="com.cornellappdev.android.eatery.MapsActivity"
             android:label="@string/title_activity_maps"
             android:theme="@style/AppTheme"
             android:screenOrientation="portrait"
@@ -58,7 +58,7 @@
 
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.example.jc.eatery_android.MainActivity" />
+                android:value="com.cornellappdev.android.eatery.MainActivity" />
         </activity>
     </application>
