+++ app/src/main/java/com/cornellappdev/android/eatery/Data/CafeteriaDbHelper.java
@@ -1,4 +1,4 @@
-package com.example.jc.eatery_android.Data;
+package com.cornellappdev.android.eatery.Data;
 
 import android.content.ContentValues;
 import android.content.Context;
