+++ app/src/main/java/com/cornellappdev/android/eatery/NetworkUtils/NetworkUtilities.java
@@ -1,6 +1,4 @@
-package com.example.jc.eatery_android.NetworkUtils;
-
-import android.util.Log;
+package com.cornellappdev.android.eatery.NetworkUtils;
 
 import java.io.BufferedReader;
 import java.io.IOException;
