+++ app/src/main/java/com/cornellappdev/android/eatery/SplashActivity.java
@@ -1,4 +1,4 @@
-package com.example.jc.eatery_android;
+package com.cornellappdev.android.eatery;
 
 import android.content.Intent;
 import android.os.Bundle;
