+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.jc.eatery_android.MainActivity"
+    tools:context="com.cornellappdev.android.eatery.MainActivity"
     android:focusableInTouchMode="true"
     >
