+++ app/src/main/res/layout/activity_maps.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.jc.eatery_android.MapsActivity"
+    tools:context="com.cornellappdev.android.eatery.MapsActivity"
     android:focusableInTouchMode="true"
     >
 
@@ -15,7 +15,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
         android:layout_marginBottom="50dp"
-    tools:context="com.example.jc.eatery_android.MapsActivity"
+    tools:context="com.cornellappdev.android.eatery.MapsActivity"
         />
